+++ packages/@shellicar/build-version/src/core/defaults.ts
@@ -1,8 +1,5 @@
-import { versionPath } from './module';
-import { DebugLevel, type Options } from './types';
+import type { Options } from './types';
 
 export const defaults = {
-  versionPath,
   debug: false,
-  debugLevel: DebugLevel.INFO,
 } satisfies Options;