+++ packages/@shellicar/build-version/src/core/index.ts
@@ -1,101 +1,34 @@
-import { execSync } from 'node:child_process';
-import type { UnpluginFactory } from 'unplugin';
-import { createUnplugin } from 'unplugin';
-import { defaults } from './defaults';
-import { createGitCalculator } from './git';
-import { createGitversionCalculator } from './gitversion';
-import { MODULE_ID } from './module';
-import { DebugLevel, type Options, type VersionCalculator } from './types';
+import { type UnpluginFactory, createUnplugin } from 'unplugin';
+import { virtualModuleId } from './module';
+import type { ILogger, Options } from './types';
+import { loadVirtualModule } from './version';
 
-const execCommand = (command: string): string => {
-  return execSync(command, { encoding: 'utf8' }).trim();
-};
+const resolveVirtualId = (id: string) => `\0${id}`;
 
-const getCalculator = (options: Options): VersionCalculator => {
-  if (typeof options?.versionCalculator === 'function') {
-    return options.versionCalculator;
-  }
-
-  switch (options.versionCalculator) {
-    case 'git':
-      return createGitCalculator({ debug: options.debug });
-    default:
-      return createGitversionCalculator(options.versionCalculator ?? 'gitversion');
-  }
-};
-
-const generateVersionInfo = (calculator: VersionCalculator) => {
-  const sha = execCommand('git rev-parse HEAD');
-  const shortSha = sha.substring(0, 7);
-  return {
-    buildDate: new Date().toISOString(),
-    branch: execCommand('git rev-parse --abbrev-ref HEAD'),
-    sha,
-    shortSha,
-    commitDate: execCommand('git log -1 --format=%cI'),
-    version: calculator(),
+const pluginFactory: UnpluginFactory<Options> = (options: Options) => {
+  const debug = options.debug ? (message?: unknown, ...optionalParams: unknown[]) => console.debug(`[version] ${message}`, ...optionalParams) : () => {};
+  const error = (message?: unknown, ...optionalParams: unknown[]) => console.error(`[version] ${message}`, ...optionalParams);
+  const logger: ILogger = {
+    debug,
+    error,
   };
-};
-
-const versionPluginFactory: UnpluginFactory<Options> = (inputOptions?: Options, meta?) => {
-  const options = {
-    ...defaults,
-    ...inputOptions,
-  };
-  const info = (message: any, ...args: any) => {
-    if (options.debug && options.debugLevel <= DebugLevel.INFO) {
-      console.info('[version] (info):', message, ...args);
-    }
-  };
-  const debug = (message: any, ...args: any) => {
-    if (options.debug && options.debugLevel <= DebugLevel.DEBUG) {
-      console.debug('[version] (dbug):', message, ...args);
-    }
-  };
-  info({ options });
-
-  const versionPattern = new RegExp(options.versionPath);
-
-  const matchVersion = (id: string) => {
-    if (meta?.framework === 'vite') {
-      return versionPattern.test(id);
-    }
-    return id === MODULE_ID;
-  };
-
-  info({ options });
-  const calculator = getCalculator(options);
 
   return {
-    name: 'version',
-    buildStart() {
-      info('Build start');
-    },
-    buildEnd() {
-      info('Build end');
-    },
-    loadInclude(id) {
-      if (matchVersion(id)) {
-        debug('loadInclude', id);
-        return true;
-      }
-    },
+    name: 'unplugin-version',
+    enforce: 'pre',
     resolveId(id) {
-      if (matchVersion(id)) {
-        debug('resoleId', id);
-        return id;
+      if (id === virtualModuleId) {
+        debug('resolveId %s', id);
+        return resolveVirtualId(virtualModuleId);
       }
     },
     load(id) {
-      if (matchVersion(id)) {
-        debug('load', id);
-        const versionInfo = generateVersionInfo(calculator);
-        const json = JSON.stringify(versionInfo, null, 2);
-        const code = `export default ${json}`;
-        return code;
+      if (id === resolveVirtualId(virtualModuleId)) {
+        debug('load %s', id);
+        return loadVirtualModule(options, logger);
       }
     },
   };
 };
 
-export const plugin = createUnplugin(versionPluginFactory);
+export const plugin = createUnplugin(pluginFactory);