+++ packages/@shellicar/build-version/src/core/module.ts
@@ -2,5 +2,4 @@ import { name } from '../../package.json';
 
 const exportSpecifier = 'version';
 
-export const MODULE_ID = `${name}/${exportSpecifier}`;
-export const versionPath = `${name}/dist/core/${exportSpecifier}\\.c?js`;
+export const virtualModuleId = `${name}/${exportSpecifier}`;